+++ Plutarch/Internal/Semigroup.hs
@@ -66,6 +66,7 @@ import Plutarch.Internal.PlutusType (
   PlutusType (PInner),
   pcon,
  )
+import Plutarch.Internal.Subtype (PSubtype, pupcast)
 import Plutarch.Internal.Term (
   S,
   Term,
@@ -130,7 +131,7 @@ instance PSemigroup PByteString where
   pstimes p bs = plet bs $ \bs' ->
     pif
       (plengthBS # bs' #== 1)
-      (preplicateBS # pto p #$ pindexBS # bs' # 0)
+      (preplicateBS # pupcast p #$ pindexBS # bs' # 0)
       (pbySquaringDefault (#<>) bs' p)
 
 {- | BLS points form a group technically, but a @PGroup@ notion would be too
@@ -413,12 +414,12 @@ instance PMonoid (PXor PByteString) where
 
 pxortimes ::
   forall (a :: S -> Type) (b :: S -> Type) (s :: S).
-  PInner a ~ PInteger =>
+  PSubtype PInteger a =>
   Term s b ->
   Term s a ->
   Term s (PXor b) ->
   Term s (PXor b)
 pxortimes def x =
   pif
-    ((pdiv # pto x # 2) #== 0)
+    ((pdiv # pupcast x # 2) #== 0)
     (pcon . PXor $ def)