+++ CHANGELOG.md
@@ -26,6 +26,8 @@
   `Plutarch.Prelude`. 
 * `pfoldlArray` renamed to `pfoldArray`.
 * `PShow`'s `pshow'` method is now available from the Prelude.
+* `PInner PPositive` is now `PNatural`, which allows the use of `pupcast` for
+  easier conversions 'upward'.
 
 ## Removed